Added answer list command
Added a command to list all the answers. I may or may not ahve forgotten my username and password for the site and wanted a quick list of answers. Obviously 5 is super wrong and I've removed whatever the hell that was and am in the process fo redoing that one.